Add files via upload
By using the datasets of Vrinda Store we tried to analyzed the data and was able to make an automated working dashboard. In this excel sheet data cleaning, data processing, analyzing and visualizing the dataset of the year 2022 to provide them with accurate results so that they can make better results for their company.